+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilteredTableModel.java
@@ -667,7 +667,7 @@ final class DefaultFilteredTableModel<R, C> extends AbstractTableModel implement
 		@Override
 		public void addListener(Runnable listener) {
 			tableModel.addDataChangedListener(listener);
-			tableModel.selectionModel().addSelectionListener(listener);
+			tableModel.selectionModel().selectionObserver().addListener(listener);
 		}
 
 		@Override