+++ swing/framework-server-monitor/src/main/java/is/codion/swing/framework/server/monitor/ui/ClientMonitorPanel.java
@@ -108,7 +108,7 @@ public final class ClientMonitorPanel extends JPanel {
 						.rightComponent(clientInstancePanel)
 						.build();
 
-		model.clientInstanceTableModel().selectionModel().addSelectedItemListener(remoteClient -> {
+		model.clientInstanceTableModel().selectionModel().selectedItemObserver().addDataListener(remoteClient -> {
 			clientInstancePanel.removeAll();
 			try {
 				if (model != null && remoteClient != null) {